moved ADVException to commons project
