+++ src/main/java/ch/hsr/array/PrimitiveArray.java
@@ -1,8 +1,8 @@
 package ch.hsr.array;
 
+import ch.hsr.adv.commons.core.logic.util.ADVException;
 import ch.hsr.adv.lib.array.logic.ArrayModule;
 import ch.hsr.adv.lib.bootstrapper.ADV;
-import ch.hsr.adv.lib.core.logic.util.ADVException;
 
 
 public class PrimitiveArray {
