+++ src/main/java/ch/hsr/array/StylesShowcase.java
@@ -2,6 +2,7 @@ package ch.hsr.array;
 
 import ch.hsr.adv.commons.core.logic.domain.styles.*;
 import ch.hsr.adv.commons.core.logic.domain.styles.presets.ADVDefaultStyle;
+import ch.hsr.adv.commons.core.logic.util.ADVException;
 import ch.hsr.adv.lib.array.logic.ArrayModule;
 import ch.hsr.adv.lib.bootstrapper.ADV;
 import ch.hsr.adv.lib.core.logic.domain.styles.ADVEnumStyle;
@@ -9,7 +10,6 @@ import ch.hsr.adv.lib.core.logic.domain.styles.presets.ADVErrorStyle;
 import ch.hsr.adv.lib.core.logic.domain.styles.presets.ADVInfoStyle;
 import ch.hsr.adv.lib.core.logic.domain.styles.presets.ADVSuccessStyle;
 import ch.hsr.adv.lib.core.logic.domain.styles.presets.ADVWarningStyle;
-import ch.hsr.adv.lib.core.logic.util.ADVException;
 
 public class StylesShowcase {
     // instantiate data structure container
