+++ src/main/java/ch/hsr/graph/NormalGraph.java
@@ -1,7 +1,7 @@
 package ch.hsr.graph;
 
+import ch.hsr.adv.commons.core.logic.util.ADVException;
 import ch.hsr.adv.lib.bootstrapper.ADV;
-import ch.hsr.adv.lib.core.logic.util.ADVException;
 import ch.hsr.adv.lib.graph.logic.GraphModule;
 import ch.hsr.graph.model.Edge;
 import ch.hsr.graph.model.Graph;
