+++ src/main/java/ch/hsr/graph/model/Edge.java
@@ -18,7 +18,7 @@ public class Edge<E> implements ADVEdge<E> {
     public Edge(Vertex sourceVertex, Vertex targetVertex, boolean directed) {
         this(sourceVertex.getId(), targetVertex.getId(), directed);
     }
-    
+
     public Edge(long sourceVertexId, long targetVertexId) {
         this(sourceVertexId, targetVertexId, false);
     }
@@ -40,8 +40,18 @@ public class Edge<E> implements ADVEdge<E> {
     }
 
     @Override
+    public void setSourceElementId(long sourceVertexId) {
+        this.sourceVertexId = sourceVertexId;
+    }
+
+    @Override
     public long getTargetElementId() {
         return targetVertexId;
+    }
+
+    @Override
+    public void setTargetElementId(long targetVertexId) {
+        this.targetVertexId = targetVertexId;
     }
 
     @Override
